Refactor app routes to a seperate module
